+++ app/src/main/java/com/eyeson/android/ui/view/main/MainViewModel.kt
@@ -3,6 +3,7 @@ package com.eyeson.android.ui.view.main
 import android.app.Application
 import android.app.Notification
 import android.content.Intent
+import android.util.Log
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import com.eyeson.sdk.EyesonAudioManager